// extensions/order-subtotal-discount/generated/input.ts
export type DiscountClass =
  /**
   * The discount is combined with an
   * [order discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  | 'ORDER'
  /**
   * The discount is combined with a
   * [product discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  | 'PRODUCT'
  /**
   * The discount is combined with a
   * [shipping discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  | 'SHIPPING';

export type CartLinesDiscountsGenerateRunInputVariables = Exact<{ [key: string]: never; }>;


export type CartLinesDiscountsGenerateRunInput = { cart: { lines: Array<{ id: string, cost: { subtotalAmount: { amount: string, currencyCode: CurrencyCode } }, merchandise:
        | { __typename: 'CustomProduct' }
        | { __typename: 'ProductVariant', id: string }
       }> }, discount: { discountClasses: Array<DiscountClass>, config: { jsonValue: unknown } | null } };